test updates for error refactor
